requirements.txt
altgraph==0.17.4
contourpy==1.3.2
CTkMessagebox==2.7
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
fonttools==4.57.0
kiwisolver==1.4.8
matplotlib==3.10.1
mpmath==1.3.0
numpy==2.2.4
packaging==24.2
pefile==2023.2.7
pillow==11.2.1
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.3
pyparsing==3.2.3
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
python-dateutil==2.9.0.post0
pywin32-ctypes==0.2.3
setuptools==78.1.0
six==1.17.0
sympy==1.13.3
